requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
openai>=1.65
httpx>=0.27,<1
shapely==2.0.6
pytest==8.3.3
pytest-mock==3.14.0
